Added ability to add custom logging options to manage log fields. For me it was needed by logstash patterns (https://github.com/elasticsearch/logstash/blob/master/patterns/haproxy)
As result option can be used like this:

    capture         => [
	'request header Host len 40',
	'request header X-Forwarded-For len 50',
	'request header Accept-Language len 50',
	'request header Referer len 200',
	'request header User-Agent len 200',
	'response header Content-Type len 30',
	'response header Content-Encoding len 10',
	'response header Cache-Control len 200',
	'response header Last-Modified len 200',
    ],

Also, if this parameter is not specified, no action with config file is performed.
I think it will be useful to add check for "option" too and do not include default value if it don't needed. For now is adds unneeded in some cases 'httpclose' option by default.